requirements.txt
langchain==0.2.13
langchain-openai==0.1.21
python-dotenv==1.0.1

# Development
rich==13.7.1